chore: move CoA support to v0.0.3
