+++ fixtures/CoA/v0.0.3/certificate.ts
