+++ fixtures/CoA/v0.0.3/template_hbs.html
@@ -539,7 +539,7 @@
   <footer style="float: left;">
     <p>
       <em class="caption">Powered by <a href="https://en10204.io">en10204.io</a></em>
-      <em style="color:lightgray; margin-left: 35%;" class="caption">https://schemas.en10204.io/coa-schemas/v0.0.2/schema.json</em>
+      <em style="color:lightgray; margin-left: 35%;" class="caption">https://schemas.en10204.io/coa-schemas/v0.0.3/schema.json</em>
       <br>
     </p>
   </footer>
