+++ fixtures/CoA/v0.0.3/valid_cert.pdf
