+++ packages/generate-coa-pdf-template/test/constants.ts
@@ -1,5 +1,5 @@
-import testCertificate from '../../../fixtures/CoA/v0.0.2/valid_cert.json';
+import testCertificate from '../../../fixtures/CoA/v0.0.3/valid_cert.json';
 
-export const defaultSchemaUrl = 'https://schemas.en10204.io/coa-schemas/v0.0.2/schema.json';
+export const defaultSchemaUrl = 'https://schemas.en10204.io/coa-schemas/v0.0.3/schema.json';
 
 export const certificate = testCertificate;
